// frontend/hooks/models/gameSessionDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GameSessionDTODjMode } from './gameSessionDTODjMode';
import type { GameSessionDTOStatus } from './gameSessionDTOStatus';
import type { PlayerDTO } from './playerDTO';
import type { RoundDTO } from './roundDTO';

export interface GameSessionDTO {
  id?: number;
  groupId?: number;
  status?: GameSessionDTOStatus;
  djMode?: GameSessionDTODjMode;
  winConditionCardCount?: number;
  currentRoundNumber?: number;
  players?: PlayerDTO[];
  currentRound?: RoundDTO;
}